requirements.txt
Django>=5.1.1
django-cors-headers
pandas>=2.1.1
scikit-learn>=1.3.2
redis>=5.0.1
numpy>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

# Spacy language model
# You'll need to run: python -m spacy download en_core_web_sm 